Recovery improvements
